+++ cmd/integration.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2024-2025 ArangoDB GmbH, Cologne, Germany
+// Copyright 2024-2026 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -23,6 +23,7 @@ package cmd
 import (
 	"github.com/spf13/cobra"
 
+	"github.com/arangodb/kube-arangodb/pkg/deployment/features"
 	integrations "github.com/arangodb/kube-arangodb/pkg/integrations"
 )
 
@@ -32,6 +33,12 @@ func init() {
 	}
 
 	if err := integrations.Register(subCommand); err != nil {
+		panic(err.Error())
+	}
+
+	// Bind Operator Feature flags so gates consumed by the integrations (e.g. SchedulerV2Workflow)
+	// are seeded from CLI flags and environment variables in the sidecar process.
+	if err := features.Init(subCommand); err != nil {
 		panic(err.Error())
 	}